+++ 1.20/Common/src/main/java/fuzs/puzzleslib/api/resources/v1/PackResourcesHelper.java
@@ -30,8 +30,42 @@ public final class PackResourcesHelper {
      * @param fixedPosition a fixed pack cannot be moved up or down, like a server or world resource pack
      * @return the {@link RepositorySource} to be added to the {@link net.minecraft.server.packs.repository.PackRepository}
      */
+    @Deprecated(forRemoval = true)
     public static RepositorySource buildClientPack(Supplier<AbstractModPackResources> factory, String id, Component title, Component description, boolean required, boolean fixedPosition) {
-        return buildClientPack(factory, new ResourceLocation(id, id), title, description, required, fixedPosition, FeatureFlagSet.of(), false);
+        return buildClientPack(factory, new ResourceLocation(id, id), title, description, required, fixedPosition);
+    }
+
+    /**
+     * Creates a new resource pack repository source (for the client).
+     * <p>Can be added via {@link fuzs.puzzleslib.api.client.core.v1.ClientModConstructor#onAddResourcePackFinders(PackRepositorySourcesContext)}.
+     *
+     * @param factory       {@link net.minecraft.server.packs.PackResources} implementation supplier
+     * @param id            id for the pack, used for internal references and is stored in <code>options.txt</code>
+     * @param title         the title of this pack shown in the pack selection screen
+     * @param description   the description for this pack shown in the pack selection screen
+     * @param required      a required pack cannot be disabled, like in the pack selection screen the pack cannot be moved to the left side; this is used for the vanilla resource pack
+     * @param fixedPosition a fixed pack cannot be moved up or down, like a server or world resource pack
+     * @return the {@link RepositorySource} to be added to the {@link net.minecraft.server.packs.repository.PackRepository}
+     */
+    public static RepositorySource buildClientPack(Supplier<AbstractModPackResources> factory, ResourceLocation id, Component title, Component description, boolean required, boolean fixedPosition) {
+        return buildClientPack(factory, id, title, description, required, fixedPosition, false);
+    }
+
+    /**
+     * Creates a new resource pack repository source (for the client).
+     * <p>Can be added via {@link fuzs.puzzleslib.api.client.core.v1.ClientModConstructor#onAddResourcePackFinders(PackRepositorySourcesContext)}.
+     *
+     * @param factory       {@link net.minecraft.server.packs.PackResources} implementation supplier
+     * @param id            id for the pack, used for internal references and is stored in <code>options.txt</code>
+     * @param title         the title of this pack shown in the pack selection screen
+     * @param description   the description for this pack shown in the pack selection screen
+     * @param required      a required pack cannot be disabled, like in the pack selection screen the pack cannot be moved to the left side; this is used for the vanilla resource pack
+     * @param fixedPosition a fixed pack cannot be moved up or down, like a server or world resource pack
+     * @param hidden        controls whether the pack is hidden from user-facing screens like the resource pack and data pack selection screens, only available on Forge
+     * @return the {@link RepositorySource} to be added to the {@link net.minecraft.server.packs.repository.PackRepository}
+     */
+    public static RepositorySource buildClientPack(Supplier<AbstractModPackResources> factory, ResourceLocation id, Component title, Component description, boolean required, boolean fixedPosition, boolean hidden) {
+        return buildClientPack(factory, id, title, description, required, fixedPosition, FeatureFlagSet.of(), hidden);
     }
 
     /**
@@ -66,8 +100,42 @@ public final class PackResourcesHelper {
      * @param fixedPosition a fixed pack cannot be moved up or down, like a server or world resource pack
      * @return the {@link RepositorySource} to be added to the {@link net.minecraft.server.packs.repository.PackRepository}
      */
+    @Deprecated(forRemoval = true)
     public static RepositorySource buildServerPack(Supplier<AbstractModPackResources> factory, String id, Component title, Component description, boolean required, boolean fixedPosition) {
-        return buildServerPack(factory, new ResourceLocation(id, id), title, description, required, fixedPosition, FeatureFlagSet.of(), false);
+        return buildServerPack(factory, new ResourceLocation(id, id), title, description, required, fixedPosition);
+    }
+
+    /**
+     * Creates a new data pack repository source (for the server).
+     * <p>Can be added via {@link fuzs.puzzleslib.api.core.v1.ModConstructor#onAddDataPackFinders(PackRepositorySourcesContext)}.
+     *
+     * @param factory       {@link net.minecraft.server.packs.PackResources} implementation supplier
+     * @param id            id for the pack, used for internal references and is stored in <code>options.txt</code>
+     * @param title         the title of this pack shown in the pack selection screen
+     * @param description   the description for this pack shown in the pack selection screen
+     * @param required      a required pack cannot be disabled, like in the pack selection screen the pack cannot be moved to the left side; this is used for the vanilla resource pack
+     * @param fixedPosition a fixed pack cannot be moved up or down, like a server or world resource pack
+     * @return the {@link RepositorySource} to be added to the {@link net.minecraft.server.packs.repository.PackRepository}
+     */
+    public static RepositorySource buildServerPack(Supplier<AbstractModPackResources> factory, ResourceLocation id, Component title, Component description, boolean required, boolean fixedPosition) {
+        return buildServerPack(factory, id, title, description, required, fixedPosition, false);
+    }
+
+    /**
+     * Creates a new data pack repository source (for the server).
+     * <p>Can be added via {@link fuzs.puzzleslib.api.core.v1.ModConstructor#onAddDataPackFinders(PackRepositorySourcesContext)}.
+     *
+     * @param factory       {@link net.minecraft.server.packs.PackResources} implementation supplier
+     * @param id            id for the pack, used for internal references and is stored in <code>options.txt</code>
+     * @param title         the title of this pack shown in the pack selection screen
+     * @param description   the description for this pack shown in the pack selection screen
+     * @param required      a required pack cannot be disabled, like in the pack selection screen the pack cannot be moved to the left side; this is used for the vanilla resource pack
+     * @param fixedPosition a fixed pack cannot be moved up or down, like a server or world resource pack
+     * @param hidden        controls whether the pack is hidden from user-facing screens like the resource pack and data pack selection screens, only available on Forge
+     * @return the {@link RepositorySource} to be added to the {@link net.minecraft.server.packs.repository.PackRepository}
+     */
+    public static RepositorySource buildServerPack(Supplier<AbstractModPackResources> factory, ResourceLocation id, Component title, Component description, boolean required, boolean fixedPosition, boolean hidden) {
+        return buildServerPack(factory, id, title, description, required, fixedPosition, FeatureFlagSet.of(), hidden);
     }
 
     /**