+++ 1.20/Common/src/main/java/fuzs/puzzleslib/impl/PuzzlesLibMod.java
@@ -2,20 +2,30 @@ package fuzs.puzzleslib.impl;
 
 import com.google.common.collect.ImmutableMap;
 import fuzs.puzzleslib.api.biome.v1.BiomeLoadingPhase;
+import fuzs.puzzleslib.api.core.v1.CommonAbstractions;
 import fuzs.puzzleslib.api.core.v1.ContentRegistrationFlags;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
 import fuzs.puzzleslib.api.core.v1.context.BiomeModificationsContext;
 import fuzs.puzzleslib.api.core.v1.context.BlockInteractionsContext;
+import fuzs.puzzleslib.api.core.v1.context.PackRepositorySourcesContext;
+import fuzs.puzzleslib.api.data.v2.AbstractRecipeProvider;
 import fuzs.puzzleslib.api.event.v1.core.EventResult;
 import fuzs.puzzleslib.api.event.v1.data.DefaultedValue;
 import fuzs.puzzleslib.api.event.v1.data.MutableInt;
 import fuzs.puzzleslib.api.event.v1.data.MutableValue;
 import fuzs.puzzleslib.api.event.v1.entity.player.GrindstoneEvents;
 import fuzs.puzzleslib.api.network.v3.NetworkHandlerV3;
+import fuzs.puzzleslib.api.resources.v1.DynamicPackResources;
+import fuzs.puzzleslib.api.resources.v1.PackResourcesHelper;
 import fuzs.puzzleslib.impl.capability.ClientboundSyncCapabilityMessage;
 import fuzs.puzzleslib.impl.core.ClientboundModListMessage;
+import fuzs.puzzleslib.impl.core.EventHandlerProvider;
 import fuzs.puzzleslib.impl.entity.ClientboundAddEntityDataMessage;
+import net.minecraft.data.recipes.FinishedRecipe;
+import net.minecraft.data.recipes.RecipeCategory;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.CommonComponents;
+import net.minecraft.network.chat.Component;
 import net.minecraft.network.protocol.game.ClientboundAddEntityPacket;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
@@ -32,6 +42,7 @@ import net.minecraft.world.level.block.Blocks;
 
 import java.util.Iterator;
 import java.util.Map;
+import java.util.function.Consumer;
 
 /**
  * This has been separated from {@link PuzzlesLib} to prevent issues with static initialization when accessing constants in {@link PuzzlesLib} early.
@@ -46,6 +57,11 @@ public class PuzzlesLibMod extends PuzzlesLib implements ModConstructor {
 
     @Override
     public void onConstructMod() {
+        EventHandlerProvider.tryRegister(CommonAbstractions.INSTANCE);
+        registerHandlers();
+    }
+
+    private static void registerHandlers() {
         Item book = Items.ENCHANTED_BOOK;
         GrindstoneEvents.UPDATE.register((ItemStack topInput, ItemStack bottomInput, MutableValue<ItemStack> output, MutableInt experienceReward, Player player) -> {
             if (topInput.isEnchanted() && bottomInput.is(book)) {
@@ -97,6 +113,35 @@ public class PuzzlesLibMod extends PuzzlesLib implements ModConstructor {
         context.registerFlattenable(Blocks.DIRT_PATH, Blocks.FARMLAND);
         context.registerTillable(Blocks.FARMLAND, Blocks.DIRT_PATH);
         context.registerTillable(Blocks.DIRT, Blocks.FARMLAND);
+    }
+
+    @Override
+    public void onAddDataPackFinders(PackRepositorySourcesContext context) {
+        context.addRepositorySource(PackResourcesHelper.buildServerPack(DynamicPackResources.create(MOD_ID, dataProviderContext -> {
+            return new AbstractRecipeProvider(dataProviderContext) {
+
+                @Override
+                public void addRecipes(Consumer<FinishedRecipe> exporter) {
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.COBBLED_DEEPSLATE, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.COBBLED_DEEPSLATE_SLAB, Blocks.DEEPSLATE, 2);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.COBBLED_DEEPSLATE_STAIRS, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.DECORATIONS, Blocks.COBBLED_DEEPSLATE_WALL, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.CHISELED_DEEPSLATE, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.POLISHED_DEEPSLATE, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.POLISHED_DEEPSLATE_SLAB, Blocks.DEEPSLATE, 2);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.POLISHED_DEEPSLATE_STAIRS, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.DECORATIONS, Blocks.POLISHED_DEEPSLATE_WALL, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.DEEPSLATE_BRICKS, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.DEEPSLATE_BRICK_SLAB, Blocks.DEEPSLATE, 2);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.DEEPSLATE_BRICK_STAIRS, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.DECORATIONS, Blocks.DEEPSLATE_BRICK_WALL, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.DEEPSLATE_TILES, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.DEEPSLATE_TILE_SLAB, Blocks.DEEPSLATE, 2);
+                    stonecutterResultFromBase(exporter, RecipeCategory.BUILDING_BLOCKS, Blocks.DEEPSLATE_TILE_STAIRS, Blocks.DEEPSLATE);
+                    stonecutterResultFromBase(exporter, RecipeCategory.DECORATIONS, Blocks.DEEPSLATE_TILE_WALL, Blocks.DEEPSLATE);
+                }
+            };
+        }), id("dynamic_recipes"), Component.literal("Dynamic Recipes"), CommonComponents.EMPTY, true, true, true));
     }
 
     @Override