+++ 1.20/Fabric/src/main/java/fuzs/puzzleslib/impl/core/FabricAbstractions.java
@@ -1,5 +1,6 @@
 package fuzs.puzzleslib.impl.core;
 
+import com.google.common.collect.Sets;
 import fuzs.puzzleslib.api.core.v1.CommonAbstractions;
 import fuzs.puzzleslib.api.event.v1.core.EventPhase;
 import fuzs.puzzleslib.api.event.v1.server.ServerLifecycleEvents;
@@ -9,6 +10,7 @@ import net.fabricmc.fabric.api.tag.convention.v1.ConventionalEntityTypeTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.packs.repository.Pack;
@@ -28,12 +30,15 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.Set;
 import java.util.function.BiConsumer;
 
-public final class FabricAbstractions implements CommonAbstractions {
+public final class FabricAbstractions implements CommonAbstractions, EventHandlerProvider {
+    private final Set<String> hiddenPacks = Sets.newConcurrentHashSet();
     private MinecraftServer gameServer;
 
-    {
+    @Override
+    public void registerHandlers() {
         // registers for game server starting and stopping, so we can keep an instance of the server here so that
         // {@link FabricNetworkHandler} can be implemented much more similarly to Forge
         ServerLifecycleEvents.STARTING.register(EventPhase.FIRST, server -> this.gameServer = server);
@@ -103,7 +108,12 @@ public final class FabricAbstractions implements CommonAbstractions {
     }
 
     @Override
-    public Pack.Info createPackInfo(Component description, int packVersion, FeatureFlagSet features, boolean hidden) {
+    public Pack.Info createPackInfo(ResourceLocation id, Component description, int packVersion, FeatureFlagSet features, boolean hidden) {
+        if (hidden) this.hiddenPacks.add(id.toString());
         return new Pack.Info(description, packVersion, features);
+    }
+
+    public boolean notHidden(String id) {
+        return !this.hiddenPacks.contains(id);
     }
 }